% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions_data.R
\name{clean_data}
\alias{clean_data}
\title{cleaning a specific variable in the data frame}
\usage{
clean_data(data, variable)
}
\arguments{
\item{data}{The data frame to be cleaned}

\item{variable}{The name of the variable to be cleaned}
}
\value{
a cleaned variable
}
\description{
cleaning a specific variable in the data frame
}
\examples{
data$age_clean <- clean_data(data, 'age');
data$bmi_clean <- clean_data(data, 'bmi');
data$fgl_clean <- clean_data(data, 'fgl');
}
