% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_function.R
\name{get_change_scale}
\alias{get_change_scale}
\title{get a colour scheme object with supplied colour scale and value ranges for a value that means change}
\usage{
get_change_scale(type, colour_ramps, minqt, maxqt, label_func)
}
\arguments{
\item{type}{type of the variable being plotted, the two options are the same in this example; they can be made different and more be added as needed}

\item{colour_ramps}{a colour ramp function with the set colour scheme}

\item{minqt}{minimum of the value being plotted}

\item{maxqt}{maximum of the value being plotted}

\item{label_func}{function to format the labels in the legend, use `waiver()` if want to pass nothing}
}
\value{
A named list with `fill` and `colour`, using the same colour scheme
}
\description{
get a colour scheme object with supplied colour scale and value ranges for a value that means change
}
