% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_function.R
\name{get_colour_scheme}
\alias{get_colour_scheme}
\title{get a colour scheme object with supplied colour scale and value ranges}
\usage{
get_colour_scheme(type, colour_ramps, scale_breaks)
}
\arguments{
\item{type}{type of the variable being plotted, a few options are provided in this example: update and add as needed}

\item{colour_ramps}{a colour ramp function with the set colour scheme}

\item{scale_breaks}{The break points for the colour scheme, starting from min and ending with max of values for the colouring variable}
}
\value{
A named list with fill and colour, using the same colour scheme
}
\description{
get a colour scheme object with supplied colour scale and value ranges
}
