% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_function.R
\name{make_map}
\alias{make_map}
\title{plot a map with default layout}
\usage{
make_map(
  data_map,
  world_map,
  my_colour_schemes,
  scale_breaks,
  type,
  plot_title = NULL,
  plot_type = "standard",
  incl_density = TRUE
)
}
\arguments{
\item{data_map}{The data frame with the value to colour the map in variable 'colour_val'}

\item{world_map}{The shape file for map}

\item{my_colour_schemes}{The colour schemes for plotting the map}

\item{scale_breaks}{The value ranges of the values being plotted}

\item{type}{The type of variable being plotted, which decides the colour scale (unless provided) and formatting of legend, title and inset density chart}

\item{plot_title}{Add plot title if provided}

\item{plot_type}{default 'standalone'; if 'large_font' then font sizes are made larger}

\item{incl_density}{whether include a density plot inside the map}
}
\value{
A ggplot grob object: can be used in egg/ggpubr ggarrange
}
\description{
plot a map with default layout
}
