% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_function.R
\name{map_function_categorical}
\alias{map_function_categorical}
\title{plot a map with categorical colour schemes and without density insert
typically used for map of number of data sources}
\usage{
map_function_categorical(
  data_map,
  world_map,
  type = "source_number",
  scale_breaks = NULL,
  colour_scale = NULL
)
}
\arguments{
\item{data_map}{The data frame with the value to colour the map in variable `colour_val`}

\item{world_map}{The shape file for map}

\item{type}{The type of variable being plotted, which decides the colour scale (unless provided) and formatting of legend and title}

\item{scale_breaks}{The cutoffs to get categories if `colour_val` is a continuous variable, decided in function if NULL}

\item{colour_scale}{Colour scheme used for plotting, matching the categories if both provided, decided in function if NULL}
}
\value{
A ggplot grob object: can be used in egg/ggpubr ggarrange
}
\description{
plot a map with categorical colour schemes and without density insert
typically used for map of number of data sources
}
\examples{
p <- map_function(data,
                  world_map,
                  c(min(data$colour_val), max(data$colour_val)))
}
